fix: Init now works as intended
Why does Cloup struct hold a value to current directory? A simple env::current_dir()
call is enough atleast when doing 'init'.
Init is only interested of the current directory so we are only asking for that.

Why have .config inside our data dir? This should reside in a config dir imo.